refactor(filetree): split monolithic filetree.rs into modules
Extract FileType enum and FileEntry struct into types.rs,
View impl into render.rs. FileTree fields changed to pub(crate)
for cross-module access. Tests remain in mod.rs (private field access).
